[FLINK-7251] [types] Remove the flink-java8 module and improve lambda type extraction
