+++ flink-examples/flink-examples-batch/src/main/java/org/apache/flink/examples/java/wordcount/WordCount.java
@@ -43,7 +43,6 @@ import org.apache.flink.util.Collector;
  * </ul>
  *
  */
-@SuppressWarnings("serial")
 public class WordCount {
 
 	// *************************************************************************
@@ -110,7 +109,7 @@ public class WordCount {
 			// emit the pairs
 			for (String token : tokens) {
 				if (token.length() > 0) {
-					out.collect(new Tuple2<String, Integer>(token, 1));
+					out.collect(new Tuple2<>(token, 1));
 				}
 			}
 		}
