+++ flink-tests/src/test/java/org/apache/flink/test/operators/ReduceITCase.java
@@ -90,7 +90,11 @@ public class ReduceITCase extends MultipleProgramsTestBase {
 
 		DataSet<Tuple5<Integer, Long, Integer, String, Long>> ds = CollectionDataSets.get5TupleDataSet(env);
 		DataSet<Tuple5<Integer, Long, Integer, String, Long>> reduceDs = ds.
-				groupBy(4, 0).reduce(new Tuple5Reduce());
+				groupBy(4, 0).reduce((in1, in2) -> {
+					Tuple5<Integer, Long, Integer, String, Long> out = new Tuple5<>();
+					out.setFields(in1.f0, in1.f1 + in2.f1, 0, "P-)", in1.f4);
+					return out;
+				});
 
 		List<Tuple5<Integer, Long, Integer, String, Long>> result = reduceDs
 				.collect();
